Move empty notarization construction to notarization.go
